added few files

+++ Maths/Theory/factorial.cpp
@@ -1,7 +1,7 @@
 #include<iostream>
 using namespace std;
 
-int fact(int n){
+long double fact(int n){
     if(n == 0)
         return 1;
     return n * fact(n - 1);
